JiraClient.get_issue handles response and returns data
