+++ mantis/jira/jira_issues.py
@@ -8,21 +8,6 @@ if TYPE_CHECKING:
     from .jira_client import JiraClient
 
 
-def process_key(key: str, exception: Exception) -> tuple[str, str]:
-    match key.split("-"):
-        case (s,):
-            raise NotImplementedError(
-                f"Partial keys are not supported. Please "
-                f'provide the full key for your issue: "PROJ-{s}"'
-            ) from exception
-        case (project, task_no):
-            return (project, task_no)
-        case _:
-            raise NotImplementedError(
-                f'Key contains too many components: "{key}"'
-            ) from exception
-
-
 class JiraIssue:
     def __init__(self, client: "JiraClient", raw_data: dict[str, Any]) -> None:
         self.client = client
@@ -89,14 +74,9 @@ class JiraIssues:
             issue_data_from_cache = self.client.cache.get_issue(key)
             if issue_data_from_cache:
                 return JiraIssue(self.client, issue_data_from_cache)
-        response = self.client.get_issue(key)
-        try:
-            response.raise_for_status()
-        except HTTPError as e:
-            self.handle_http_error(e, key)
-        issue_data: dict[str, dict] = response.json()
-        self.client.cache.write_issue(key, issue_data)
-        return JiraIssue(self.client, issue_data)
+        data = self.client.get_issue(key)
+        self.client.cache.write_issue(key, data)
+        return JiraIssue(self.client, data)
 
     def create(self, issuetype: str, title: str, data: dict) -> dict:
         assert issuetype in self.allowed_types
@@ -111,29 +91,3 @@ class JiraIssues:
         response.raise_for_status()
         response_data: dict = response.json()
         return response_data
-
-    def handle_http_error(self, exception: HTTPError, key: str) -> None:
-        (project_from_key, task_no_from_key) = process_key(key, exception)
-        match exception.response.reason:
-            case "Not Found":
-                if " " in key:
-                    raise ValueError(
-                        f'Whitespace in key is not allowed ("{key}")'
-                    ) from exception
-                elif not task_no_from_key.isnumeric():
-                    raise ValueError(
-                        f'Issue number "{task_no_from_key}" in key "{key}" must be numeric'
-                    ) from exception
-                elif self.client.options.project not in key:
-                    raise ValueError(
-                        f"The requested issue does not exist. Note that the "
-                        f'provided key "{key}" does not appear to match '
-                        f'your configured project "{self.client.options.project}"'
-                    ) from exception
-                else:
-                    raise ValueError(
-                        f'The issue "{project_from_key}-{task_no_from_key}" does '
-                        f'not exists in the project "{project_from_key}"'
-                    ) from exception
-            case _:
-                raise AttributeError("Unknown reason") from exception
